# app/static/swagger.yaml
...
openapi: 3.0.0
info:
  description:
    REST API built using Flask to allow users to track the board games
    they play, share reviews with others and add games to their collection.
  version: 1.0.0
  title: Board Game Tracker
...
paths:
...
  "/api/reviews":
    get:
      tags:
        - Reviews
      summary: Retrieve all reviews
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/components/schemas/reviews"
        "404":
          description: No reviews found
    post:
      tags:
        - Reviews
      summary: Add a new review
      requestBody:
        description: Review Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/reviewRequestPostBody"
      produces:
        - application/json
      responses:
        "201":
          description: OK
          schema:
            "$ref": "#/components/schemas/reviewPostResponse"
        "401":
          description: Authentication error
        "403":
          description: Authorization error
        "422":
          description: Failed. Bad post data.
  "/api/reviews/{id}":
    get:
      tags:
        - Reviews
      summary: Returns review by ID
      parameters:
        - name: id
          in: path
          description: ID of review to retrieve
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/components/schemas/reviews/id"
        "404":
          description: Invalid id supplied
    patch:
      tags:
        - Reviews
      summary: Update a review
      requestBody:
        description: Review Request Patch Object
        required: false
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/genreRequestPatchBody"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            "$ref": "#/components/schemas/reviewPatchResponse"
        "401":
          description: Authentication error
        "403":
          description: Authorization error
        "404":
          description: Invalid id supplied
        "422":
          description: Failed. Bad post data.
  "/api/reviews/{id}/reactions":
    get:
      tags:
        - Reviews
      summary: Returns likes and dislikes for a specific review
      parameters:
        - name: id
          in: path
          description: ID of review to retrieve reactions for
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/components/schemas/reviews/id/reactions"
        "404":
          description: Invalid id supplied
    patch:
      tags:
        - Reviews
      summary: React to a review (like or dislike)
      requestBody:
        description: Reaction Request Patch Object
        required: false
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/reactionRequestPatchBody/"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            "$ref": "#/components/schemas/reactionPatchResponse"
        "401":
          description: Authentication error
        "403":
          description: Authorization error
        "404":
          description: Invalid id supplied
        "422":
          description: Failed. Bad post data.

  # BOARD GAMES
...